+++ CHANGELOG.rst
@@ -2,8 +2,29 @@
 Changelog
 =========
 
+0.x.x (?)
+==================
+
+* Added ...
+* Added Minimum option for Timeseries
+* Added Maximum option for Timeseries
+* Added Number of decimals displays option for Timeseries
+* Added Bar_Chart_ panel support
+* Extended SqlTarget to support parsing queries from files
+* Fix AlertCondition backwards compatibility (``useNewAlerts`` default to ``False``)
+* Added RateMetricAgg_ for ElasticSearch
+* added axisSoftMin and axisSoftMax options for TimeSeries
+* Added support for Azure Data Explorer datasource plugin (https://github.com/grafana/azure-data-explorer-datasource)
+
+.. _`Bar_Chart`: https://grafana.com/docs/grafana/latest/panels-visualizations/visualizations/bar-chart/
+.. _`RateMetricAgg`: https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-metrics-rate-aggregation.html
+
+* Added unit parameter to the Table class in core
+* Added a hide parameter to ElasticsearchTarget
+* Fix value literal GAUGE_CALC_TOTAL to sum instead of total
+
 0.7.0 (2022-10-02)
-===========
+==================
 
 * Added Grafana 8.x new Alert Rule
 * Added Grafana 9.x new Alert Rule
@@ -15,8 +36,7 @@ Changelog
 * Fix typo in unit constant ``GIGA_WATT`` (was ``GAGA_WATT``)
 * Fix typo in unit constant ``NORMAL_CUBIC_METER`` (was ``NORMAIL_CUBIC_METER``)
 
-.. _ePict: basehttps://grafana.com/grafana/plugins/larona-epict-panel/
-
+.. _`ePict`: https://grafana.com/grafana/plugins/larona-epict-panel/
 
 0.6.3 (2022-03-30)
 ==================